+++ backend/src/app.module.ts
@@ -1,0 +1,11 @@
+import { Module } from "@nestjs/common";
+import { AppController } from "./app/controller/app.controller";
+import { AppService } from "./app/service/app.service";
+import { UsersModule } from "./users/users.module";
+
+@Module({
+  imports: [UsersModule],
+  controllers: [AppController],
+  providers: [AppService],
+})
+export class AppModule { }
