--- backend/src/app/app.module.ts
@@ -1,11 +1,0 @@
-import { Module } from "@nestjs/common";
-import { AppController } from "./app.controller";
-import { AppService } from "./app.service";
-import { UsersModule } from "../users/users.module";
-
-@Module({
-  imports: [UsersModule],
-  controllers: [AppController],
-  providers: [AppService],
-})
-export class AppModule { }
