+++ backend/src/app/entity/apphttpexception.entity.ts
