+++ backend/src/users/controller/users.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { UsersService } from "./users.service";
 import { UsersController } from "./users.controller";
+import { UsersService } from "users/service/users.service";
